Name : Mitul TaskID:  546 Module : CTMS SubModule :  Activity Master for pass through cost Description : 1)Grid should contain Activity| STudy, site,patient| Unit(visit, month,year etc..)| Rate| Frequency|No of unit| Amount      (Rate * Frequency* No of Unit= Total Amount)

+++ GSC.Data/Dto/CTMS/StudyPlanTaskDto.cs
@@ -52,7 +52,7 @@ namespace GSC.Data.Dto.CTMS
         public string StudayName { get; set; }
         public string SiteName { get; set; }
         public string CountryName { get; set; }
-        public int? TotalCost { get; set; }
+        public decimal? TotalCost { get; set; }
         public List<StudyPlanTaskDto> Subtasks { get; set; }
         public List<ResourceTypeGridDto> TaskResource { get; set; }
     }
